refactor: move deno json functionality to args module (#22710)
