+++ cli/tools/registry/mod.rs
@@ -59,7 +59,6 @@ use auth::get_auth_method;
 use auth::AuthMethod;
 pub use pm::add;
 use publish_order::PublishOrderGraph;
-pub use unfurl::deno_json_deps;
 use unfurl::SpecifierUnfurler;
 
 use super::check::TypeChecker;
